requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
faiss-cpu==1.11.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
ftfy==6.3.1
grpcio==1.67.1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multilingual-clip==1.0.10
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
open_clip_torch==2.32.0
openai==1.82.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.0
pydantic==2.11.5
pydantic_core==2.33.2
pymilvus==2.5.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
wcwidth==0.2.13
wheel==0.45.1
